% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{diff_fun}
\alias{diff_fun}
\title{Difference function}
\usage{
diff_fun(input)
}
\arguments{
\item{input}{A vector to find difference.}
}
\value{
Calculated difference vector.
}
\description{
Find difference for the provided input vector.
}
