% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{lag_fun}
\alias{lag_fun}
\title{Lagging function}
\usage{
lag_fun(input, order_lag = 1)
}
\arguments{
\item{input}{A vector to find lag.}

\item{order_lag}{A double as lag order.}
}
\value{
Calculated lag vector.
}
\description{
Find lag of \code{order_lag} for the provided input vector.
}
