% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{lead_fun}
\alias{lead_fun}
\title{Leading function}
\usage{
lead_fun(input, order_lead = 1)
}
\arguments{
\item{input}{A vector to find lead.}

\item{order_lead}{A double as lead order.}
}
\value{
Calculated lead vector.
}
\description{
Find lead of \code{order_lead} for the provided input vector.
}
