% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mian_function.R
\name{main_fun}
\alias{main_fun}
\title{Main function}
\usage{
main_fun(input_data = default_input_data)
}
\arguments{
\item{input_data}{A data frame as raw input.}
}
\value{
A list containing a list of meta data and a data frame including all
the raw input and new transformations.
}
\description{
Interactive function to ask users for their coice of input and
transformations.
}
