% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{prod_fun}
\alias{prod_fun}
\title{Product function}
\usage{
prod_fun(input_1, input_2)
}
\arguments{
\item{input_1}{A vector as the first input.}

\item{input_2}{A vector as the second input.}
}
\value{
Calculated product vector.
}
\description{
Find product of the provided input vectors.
}
