% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{ratio_fun}
\alias{ratio_fun}
\title{Ratio function}
\usage{
ratio_fun(input_1, input_2)
}
\arguments{
\item{input_1}{A vector as the first input.}

\item{input_2}{A vector as the second input.}
}
\value{
Calculated ratio vector.
}
\description{
Find ratio between the provided input vectors.
}
