% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{rolling_mean}
\alias{rolling_mean}
\title{Rolling mean function}
\usage{
rolling_mean(input, win = 2)
}
\arguments{
\item{input}{A vector to find rolling mean.}

\item{win}{A double as rolling window.}
}
\value{
Calculated rolling mean column.
}
\description{
Find rolling mean for the provided input and window.
}
