% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{rolling_sd}
\alias{rolling_sd}
\title{Rolling sd function}
\usage{
rolling_sd(input, win = 2)
}
\arguments{
\item{input}{A vector to find rolling sd.}

\item{win}{A double as rolling window.}
}
\value{
Calculated rolling sd column.
}
\description{
Find rolling sd for the provided input and window.
}
