% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{spread_fun}
\alias{spread_fun}
\title{Spread function}
\usage{
spread_fun(input_1, input_2)
}
\arguments{
\item{input_1}{A vector as the first input.}

\item{input_2}{A vector as the second input.}
}
\value{
Calculated spread vector.
}
\description{
Find spread between the provided input vectors.
}
